Here are the changes I did during class time. (Feb 12, 2025). Code from today in class

+++ app/src/main/java/com/example/earthquakelistproject/Properties.kt
@@ -5,6 +5,6 @@ data class Properties(
     var mag: Double,
     var title: String,
     var place: String,
-    var time: Int,
+    var time: Long,
     var url: String
 )
